Refactor parallelization into separate classes
